Refactored game end conditions into game_logic
